lxd/identity: Moves identity methods from shared/auth to lxd/identity.
These methods are only required by the LXD server.
